HTTP Post/Get key retrieving
Add $inline parameter into key retrieving methods to determine if a key should be inline trimed or not (can be replaced by a type checking, but not sure it worth it because sometimes a simple text or html code can be inline too)

Add $type parameter into secureUnknownType

Add type managment into secureUnknownType as 'htlm' type need a special process

Add strip_tags() method to securise data that are not html data

Force addslashes() use instead of SqlSecureMethod call into secure methods. We don't need SqlSecureMethod, as we don't know what it does (it's part of SGBDR) and as securing http data is not part of SQL process as we do not use that data into a SQL treatment for now. SQL protect/quote should only be used with SQL treatment. It's bfw-sql (or other sql module) purpose.